requirements.txt
ansi2html==1.9.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.8
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mutagen==1.47.0
nest-asyncio==1.6.0
numpy @ file:///home/builder/.termux-build/python-numpy/src
packaging==23.2
pandas @ file:///home/builder/.termux-build/python-pandas/src/dist/pandas-2.2.0-cp311-cp311-linux_aarch64.whl
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
